Fix some of the calculators
Particularly binary, accordance and congeneric (but not completely).

+++ LibiadaWeb/Models/Repositories/Catalogs/CongenericCharacteristicRepository.cs
@@ -168,12 +168,12 @@
             {
                 List<SelectListItem> linkSelectListItems = characteristicsLinks
                     .Where(cl => cl.CongenericCharacteristic == characteristic && links.Contains(cl.Link))
-                    .Select(cl => new SelectListItem { Value = cl.Link.ToString(), Text = cl.Link.GetDisplayValue() })
+                    .Select(cl => new SelectListItem { Value = ((byte)cl.Link).ToString(), Text = cl.Link.GetDisplayValue() })
                     .Distinct(new SelectListItemComparer())
                     .ToList();
                 List<SelectListItem> arrangementTypeSelectListItems = characteristicsLinks
                     .Where(cl => cl.CongenericCharacteristic == characteristic && arrangementTypes.Contains(cl.ArrangementType))
-                    .Select(cl => new SelectListItem { Value = cl.ArrangementType.ToString(), Text = cl.ArrangementType.GetDisplayValue() })
+                    .Select(cl => new SelectListItem { Value = ((byte)cl.ArrangementType).ToString(), Text = cl.ArrangementType.GetDisplayValue() })
                     .Distinct(new SelectListItemComparer())
                     .ToList();
